refactor: cart quantity button
