+++ src/advanced/components/index.ts
@@ -1,4 +1,4 @@
-export * from './cart-item-container';
+export * from './cart/container';
 export * from './product-select';
 export * from './stock-status-display';
 export * from './total-amount-display';
